+++ R/tar_progress_summary.R
@@ -59,7 +59,7 @@ tar_progress_summary <- function(
     completed = sum(progress$progress == "completed"),
     errored = sum(progress$progress == "errored"),
     canceled = sum(progress$progress == "canceled"),
-    since = units_seconds(difftime(Sys.time(), time, units = "secs")),
+    since = prettyunits::pretty_sec(difftime(Sys.time(), time, units = "secs")),
     time = time_stamp(time)
   )
   fields_quosure <- rlang::enquo(fields)