+++ R/utils_cli.R
@@ -5,7 +5,7 @@ cli_pipeline_uptodate <- function(
   time <- if_any(time_stamp, time_stamp_cli(), NULL)
   msg <- paste(c(time, "skipped pipeline"), collapse = " ")
   if (!is.null(seconds_elapsed)) {
-    msg_time <- paste0(" [", units_seconds(seconds_elapsed), "]")
+    msg_time <- paste0(" [", prettyunits::pretty_sec(seconds_elapsed), "]")
     msg <- paste0(msg, msg_time)
   }
   cli::cli_alert_success(msg)
@@ -22,7 +22,7 @@ cli_pipeline_done <- function(
   if (!is.null(seconds_elapsed)) {
     msg_time <- sprintf(
       " [%s, %s completed, %s skipped]",
-      units_seconds(seconds_elapsed),
+      prettyunits::pretty_sec(seconds_elapsed),
       completed,
       skipped
     )
@@ -38,7 +38,7 @@ cli_pipeline_empty <- function(
   time <- if_any(time_stamp, time_stamp_cli(), NULL)
   msg <- paste(c(time, "empty pipeline"), collapse = " ")
   if (!is.null(seconds_elapsed)) {
-    msg_time <- paste0(" [", units_seconds(seconds_elapsed), "]")
+    msg_time <- paste0(" [", prettyunits::pretty_sec(seconds_elapsed), "]")
     msg <- paste0(msg, msg_time)
   }
   cli::cli_alert_warning(msg)
@@ -51,7 +51,7 @@ cli_pipeline_errored <- function(
   time <- if_any(time_stamp, time_stamp_cli(), NULL)
   msg <- paste(c(time, "errored pipeline"), collapse = " ")
   if (!is.null(seconds_elapsed)) {
-    msg_time <- paste0(" [", units_seconds(seconds_elapsed), "]")
+    msg_time <- paste0(" [", prettyunits::pretty_sec(seconds_elapsed), "]")
     msg <- paste0(msg, msg_time)
   }
   cli::cli_alert_danger(msg)
@@ -72,10 +72,10 @@ cli_resources <- function(target) {
     target$patternview$bytes
   metrics <- NULL
   if (!is.null(seconds_elapsed) && !anyNA(seconds_elapsed)) {
-    metrics <- c(metrics, units_seconds(seconds_elapsed))
+    metrics <- c(metrics, prettyunits::pretty_sec(seconds_elapsed))
   }
   if (!is.null(bytes_storage) && !anyNA(bytes_storage)) {
-    metrics <- c(metrics, units_bytes(bytes_storage))
+    metrics <- c(metrics, prettyunits::pretty_bytes(bytes_storage))
   }
   if (!is.null(metrics)) {
     metrics <- paste0("[", paste(metrics, collapse = ", "), "]")